refactor(node-renderer): simplify selectors, remove reselect, cleanup
