+++ packages/react/package.json
@@ -57,7 +57,6 @@
     "d3-drag": "^3.0.0",
     "d3-selection": "^3.0.0",
     "d3-zoom": "^3.0.0",
-    "reselect": "^4.1.8",
     "zustand": "^4.4.0"
   },
   "peerDependencies": {
