+++ packages/react/src/types/nodes.ts
@@ -1,5 +1,5 @@
 import type { CSSProperties, MouseEvent as ReactMouseEvent } from 'react';
-import type { NodeBase, XYPosition } from '@xyflow/system';
+import type { NodeBase } from '@xyflow/system';
 
 export type Node<NodeData = any, NodeType extends string | undefined = string | undefined> = NodeBase<
   NodeData,
@@ -20,11 +20,10 @@ export type WrapNodeProps<NodeData = any> = Pick<
 > &
   Required<Pick<Node<NodeData>, 'selected' | 'type' | 'zIndex'>> & {
     isConnectable: boolean;
-    xPos: number;
-    yPos: number;
     xPosOrigin: number;
     yPosOrigin: number;
-    positionAbsolute: XYPosition;
+    positionAbsoluteX: number;
+    positionAbsoluteY: number;
     initialized: boolean;
     isSelectable: boolean;
     isDraggable: boolean;
